Refactor : Service layer class로 수정
