+++ src/main/java/com/prgrms/jpa/service/UserService.java
@@ -2,8 +2,26 @@ package com.prgrms.jpa.service;
 
 import com.prgrms.jpa.controller.dto.user.CreateUserRequest;
 import com.prgrms.jpa.controller.dto.user.UserIdResponse;
+import com.prgrms.jpa.domain.User;
+import com.prgrms.jpa.repository.UserRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface UserService {
+import static com.prgrms.jpa.utils.ToDtoMapper.toUserIdDto;
+import static com.prgrms.jpa.utils.ToEntityMapper.toUser;
 
-    UserIdResponse create(CreateUserRequest createUserRequest);
+@Service
+public class UserService{
+
+    private final UserRepository userRepository;
+
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Transactional
+    public UserIdResponse create(CreateUserRequest createUserRequest) {
+        User user = userRepository.save(toUser(createUserRequest));
+        return toUserIdDto(user.getId());
+    }
 }
